requirements.txt
Pillow
requests
re
json
time
logging
math
duckduckgo-images-api
duckduckpy